improve: ADMIN 권한의 API 패키지 통합
