+++ src/main/java/com/techfork/global/constant/Constants.java
@@ -26,7 +26,6 @@ public final class Constants {
     };
 
     public static final String[] ADMIN_ENDPOINTS = {
-            "/api/v1/batch/**",
             "/api/v1/admin/**"
     };
